% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeltaLE.R
\name{getDeltaLE}
\alias{getDeltaLE}
\title{getDeltaLE}
\usage{
getDeltaLE(gap_country)
}
\arguments{
\item{gap_country}{char}
}
\value{
numeric
}
\description{
calculates the Delta of LE from 2007 to 1952 for a given country
}
\examples{
\dontrun{
getDeltaLE("France")
}
}
