% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printLE.R
\name{printLE}
\alias{printLE}
\title{printLE}
\usage{
printLE(gap_country, gap_year)
}
\arguments{
\item{gap_country}{character}

\item{gap_year}{integer}
}
\value{
character
}
\description{
Given year and country, returns a statement for Life exp from gapminder ds
}
\examples{
\dontrun{
printLE("Afghanistan", 2007)
}
}
