% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCountry.R
\name{plotCountry}
\alias{plotCountry}
\title{plotCountry
plot the life expectancy vs GDP for a given country}
\usage{
plotCountry(name)
}
\arguments{
\item{name}{country of gapminder}
}
\value{
plot
}
\description{
plotCountry
plot the life expectancy vs GDP for a given country
}
\examples{
' @examples
'\dontrun{
plotCountry("Germany")
}
}
